Extract topic range value helpers
